### 2026-05-03 #10504 @arfon
@zawadzkim – please merge your code (main) branch into `joss` so we can use @editorialbot to check the repository
