### 2026-07-07 4b87e20
paper: fixes in the paper and bib file
- an addiontional citation for flopy was added
- DOIs for a few references were added
- fixed the AI Disclosure section by removing accidental indentation.
