+++ paper/paper.bib
@@ -115,6 +115,18 @@
 	year = {2024}
 }
 
+@article{bakker_flopy_2016,
+	author = {Bakker, Mark and Post, Vincent and Langevin, Christian D. and Hughes, Joseph D. and White, Jeremy T. and Starn, J. Jeffrey and Fienen, Michael N.},
+	title = {Scripting {MODFLOW} Model Development Using {Python} and {FloPy}},
+	journal = {Groundwater},
+	volume = {54},
+	number = {5},
+	pages = {733--739},
+	doi = {10.1111/gwat.12413},
+	url = {https://ngwa.onlinelibrary.wiley.com/doi/abs/10.1111/gwat.12413},
+	year = {2016}
+}
+
 @misc{dewit_wofost_crop_parameters,
   author       = {de Wit, Allard and Burggraaff, Olivier},
   title        = {{WOFOST\_crop\_parameters: Parameter sets for the WOFOST crop simulation model implemented in YAML}},
@@ -125,12 +137,12 @@
 }
 
 @misc{vandenbrink_dutchsoils,
-  author       = {van den Brink, Mark},
-  title        = {{DutchSoils: Python package to get soil data from the Dutch Soil Map, Staring series and BOFEK clustering}},
-  year         = {2025},
-  publisher    = {GitHub},
-  howpublished = {\url{https://github.com/markvdbrink/dutchsoils}},
-  note         = {Accessed: November 22, 2025}
+  author = {van den Brink, Mark},
+  doi = {10.5281/zenodo.21106157},
+  month = {5},
+  title = {DutchSoils: Python package to get Dutch soil data},
+  url = {https://dutchsoils.readthedocs.io},
+  year = {2025}
 }
 
 @misc{zawadzki_dovwms,
@@ -179,7 +191,8 @@
   author={Zawadzki, Mateusz},
   booktitle={EGU General Assembly Conference Abstracts},
   pages={19745},
-  year={2024}
+  year={2024},
+  doi={10.5194/egusphere-egu24-19745}
 }
 
 @article{hughesMODFLOWApplicationProgramming2022,
