+++ paper/paper.md
@@ -48,11 +48,11 @@ Like many Fortran programs, SWAP requires a set of plain-text ASCII files with c
 
 The only published tools for SWAP input/output handling are currently written in R (e.g., rSWAP [@moritzshore_rswap_2023] and SWAPTools, distributed with SWAP 4.2.0). Prior to pySWAP, no structured, tested, and peer-reviewed Python library for SWAP model interaction existed. Given that Python has become the dominant language for scientific data analysis and hydrological modeling workflows, this gap limits the accessibility of SWAP and therefore the potential for community-driven model development.
 
-The precedent for wrapping Fortran-based hydrological models in Python is well established — most notably by flopy [@bakker_flopy_2025; @hughes_flopy_2024], which demonstrated that an object-oriented Python interface to MODFLOW significantly lowers barriers to automation, reproducibility, and community adoption. Beyond file handling, a native Python interface unlocks the full scientific ecosystem — from statistical analysis and machine learning to visualization and model coupling — enabling modelers to integrate SWAP directly into their broader research workflows without manual data exchange [@hughesMODFLOWApplicationProgramming2022]. The pySWAP architecture is also designed to anticipate the evolution of SWAP itself: as the SWAP fortran codebase is currently maturing toward a callable library interface, tools such as f2py will allow direct in-memory coupling, eliminating file-based execution entirely.
+The precedent for wrapping Fortran-based hydrological models in Python is well established — most notably by flopy [@bakker_flopy_2016; @bakker_flopy_2025; @hughes_flopy_2024], which demonstrated that an object-oriented Python interface to MODFLOW significantly lowers barriers to automation, reproducibility, and community adoption. Beyond file handling, a native Python interface unlocks the full scientific ecosystem — from statistical analysis and machine learning to visualization and model coupling — enabling modelers to integrate SWAP directly into their broader research workflows without manual data exchange [@hughesMODFLOWApplicationProgramming2022]. The pySWAP architecture is also designed to anticipate the evolution of SWAP itself: as the SWAP fortran codebase is currently maturing toward a callable library interface, tools such as f2py will allow direct in-memory coupling, eliminating file-based execution entirely.
 
 # Software design
 
-pySWAP's object-oriented design is inspired by `flopy` [@bakker_flopy_2025; @hughes_flopy_2024], with functionality split into modules for intuitive access. The design is guided by three principles: (1) map SWAP's conceptual structure directly onto Python objects so the model is readable and auditable without consulting raw input files, (2) leverage the Python scientific ecosystem (numpy, pandas, pydantic) rather than reimplementing standard functionality, and (3) maintain separation between model definition, execution, and result analysis for easier extensibility.
+pySWAP's object-oriented design is inspired by `flopy` [@bakker_flopy_2016; @bakker_flopy_2025; @hughes_flopy_2024], with functionality split into modules for intuitive access. The design is guided by three principles: (1) map SWAP's conceptual structure directly onto Python objects so the model is readable and auditable without consulting raw input files, (2) leverage the Python scientific ecosystem (numpy, pandas, pydantic) rather than reimplementing standard functionality, and (3) maintain separation between model definition, execution, and result analysis for easier extensibility.
 
 Users primarily interact with the `components` and `model` modules. Each component (e.g., `meteorology`, `crop`) groups co-dependent attributes corresponding to sections of the original SWAP input files, with input validation handled via `pydantic` to catch incorrect model setups before any files are written or the model is executed. Once assembled into a `Model` object, these components build input files directory, call the SWAP executable, and return a structured `Result` object used in further analysis.
 
@@ -129,7 +129,8 @@ The package implements robust continuous integration and deployment. Package and
 
[…]
