### 2026-08-20 #11176 @editorialbot
[REVIEW]: Bracing the SWAP hydrological model for the future with the pySWAP Python package
**Submitting author:** @zawadzkim (<a href="http://orcid.org/0000-0001-9524-4208">Mateusz Zawadzki</a>)
**Repository:** https://github.com/zawadzkim/pySWAP
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.9
