### 2026-08-25 #11176 @zawadzkim
Thank you @jlarsen-usgs and @martinvonk for agreeing to review this package! For clarification, branch `main` is the one used for the build; branch joss contains all the commits from main plus the paper, so pyswap should install correctly from both branches.
