### 2026-08-29 222776c
tests: add SSDI serialization tests
