### 2026-08-29 f9ec29a
feat: add SSDI (subsurface drip irrigation) support
Expose SWAP 4.2.0's SSDI source term (SWSSDI switch + separate events
file) through FixedIrrigation: SSDIEVENTS table, application depth
ssdi_z, serialization of the .swp switch block and the SSDI file, and
staging of the file by the model builder. Fixed-events mode only.
