+++ pyswap/components/irrigation.py
@@ -18,13 +18,15 @@ from typing import Literal as _Literal
 from pydantic import (
     Field as _Field,
     PrivateAttr as _PrivateAttr,
+    model_validator as _model_validator,
 )
 
-from pyswap.components.tables import IRRIGEVENTS
+from pyswap.components.tables import IRRIGEVENTS, SSDIEVENTS
 from pyswap.core.basemodel import PySWAPBaseModel as _PySWAPBaseModel
 from pyswap.core.defaults import FNAME_IN as _FNAME_IN
 from pyswap.core.fields import (
     DayMonth as _DayMonth,
+    Decimal2f as _Decimal2f,
     String as _String,
     Table as _Table,
 )
@@ -35,7 +37,7 @@ from pyswap.utils.mixins import (
     YAMLValidatorMixin as _YAMLValidatorMixin,
 )
 
-__all__ = ["IRRIGEVENTS", "FixedIrrigation", "ScheduledIrrigation"]
+__all__ = ["IRRIGEVENTS", "SSDIEVENTS", "FixedIrrigation", "ScheduledIrrigation"]
 
 
 class FixedIrrigation(
@@ -48,6 +50,16 @@ class FixedIrrigation(
         swirgfil (Literal[0, 1]): Switch for separate file with fixed irrigation applications
         irrigevents (Optional[Table]):
         irgfil (Optional[str]):
+        swssdi (Optional[Literal[0, 1]]): Switch for subsurface drip irrigation
+            (SSDI) — a source term at depth in the Richards equation
+            (SWAP >= 4.2.0; documented in the SWAP manual section 11.2.4).
+        ssdi_file (Optional[str]): Name of the SSDI input file, written next to
+            the .swp file (default 'swap.ssdi').
+        ssdi_schedule (Optional[Literal[0]]): SSDI scheduling mode; only the
+            fixed-events mode (0) is currently supported.
+        ssdievents (Optional[Table]): SSDIEVENTS table with the dated
+            applications (date, rate [mm/h], amount [mm]).
+        ssdi_z (Optional[float]): Application depth [-100..0 cm].
     """
 
     _extension = _PrivateAttr(default="irg")
@@ -56,21 +68,61 @@ class FixedIrrigation(
     swirgfil: _Literal[0, 1] | None = None
     irgfil: _String = _Field(default=_FNAME_IN, frozen=True)
     irrigevents: _Table | None = None
+    swssdi: _Literal[0, 1] | None = None
+    ssdi_file: _String | None = None
+    ssdi_schedule: _Literal[0] | None = None
+    ssdievents: _Table | None = None
+    ssdi_z: _Decimal2f | None = _Field(default=None, ge=-100.0, le=0.0)
+
+    # Fields that belong to the separate SSDI file, not to the .swp section.
+    _ssdi_content = {"ssdi_schedule", "ssdievents", "ssdi_z"}
+
+    @_model_validator(mode="after")
+    def _ssdi_defaults(self):
+        if self.swssdi == 1:
+            if self.ssdi_file is None:
+                self.ssdi_file = "swap.ssdi"
+            if self.ssdi_schedule is None:
+                self.ssdi_schedule = 0
+        return self
 
     def model_string(self, **kwargs) -> str:
         """Override the model_string to handle optional file generation.
 
         Return the full section if swirgfil is set to 1, otherwise, irrigevents
         is excluded from the string and saved in a separate .irg file.
+        The SSDI content fields always go to the separate SSDI file; only the
+        SWSSDI switch and the file name appear in the .swp section.
         """
+        exclude = set(self._ssdi_content)
         if self.swirgfil == 1:
-            return super().model_string(exclude={"irrigevents"}, **kwargs)
-        else:
-            return super().model_string()
+            exclude.add("irrigevents")
[…]
