### 2026-08-29 bcba6ea
feat: add declarative parameter core for sensitivity and estimation
Param/ParameterSet address locations in a Model by dotted path
(component attributes and table cells), apply values through the
existing update machinery, and carry bounds and sampling transforms.
Observation/Observations pair observed series with user-supplied
extractors of their simulated counterparts and produce weighted
residual vectors. evaluate() runs a batch of samples via run_parallel.
Shared core for the upcoming pyswap.sensitivity and pyswap.estimation
layers; compound parameters (one value, several locations) are ou
[…]
