+++ pyswap/parameters.py
@@ -1,0 +1,238 @@
+"""Declarative parameter addressing for sensitivity analysis and estimation.
+
+This module is the shared core of the model-analysis tooling: it defines how a
+named parameter maps onto a location inside a `Model`, how a set of parameter
+values is applied to produce a new model, and how observed series are compared
+against simulation results.
+
+The sensitivity (`pyswap.sensitivity`) and estimation (`pyswap.estimation`)
+layers consume these objects; they contain no addressing logic of their own.
+
+Paths address two kinds of locations:
+
+- component attributes, dotted from the model root, e.g.
+  ``"evaporation.rsoil"`` or ``"lateraldrainage.drafile.fluxes.drares1"``;
+- table cells, with the column name and integer row index as the final
+  segment, e.g. ``"soilprofile.soilhydrfunc.KSATFIT[2]"``.
+
+Compound parameters (one value driving several locations) are not expressible
+as a single path; list one parameter per location instead.
+
+Classes:
+    Param: One named, bounded parameter addressing a location in the model.
+    ParameterSet: An ordered collection of parameters.
+    Observation: An observed series paired with an extractor of its simulated
+        counterpart.
+    Observations: A collection of observations producing residual vectors.
+
+Functions:
+    evaluate: Run the model for a batch of parameter samples.
+"""
+
+from __future__ import annotations
+
+import math
+import re
+from collections.abc import Callable
+from dataclasses import dataclass, field
+from typing import TYPE_CHECKING, Literal
+
+import numpy as np
+import pandas as pd
+
+if TYPE_CHECKING:
+    from pyswap.model.model import Model
+    from pyswap.model.result import Result
+
+__all__ = ["Observation", "Observations", "Param", "ParameterSet", "evaluate"]
+
+_CELL_RE = re.compile(r"^([A-Za-z_][A-Za-z0-9_]*)\[(\d+)\]$")
+
+
+@dataclass(frozen=True)
+class Param:
+    """One named, bounded parameter addressing a location in the model.
+
+    Attributes:
+        name: Identifier used in samples, results and reports.
+        path: Dotted path from the model root (see module docstring).
+        bounds: Lower and upper bound in physical (untransformed) units.
+        transform: Optional sampling-space transform; samplers and optimizers
+            work in transformed space, `ParameterSet.apply` always takes
+            physical values.
+    """
+
+    name: str
+    path: str
+    bounds: tuple[float, float]
+    transform: Literal["log10"] | None = None
+
+    def to_sample_space(self, value: float) -> float:
+        """Physical value -> sampling/optimization space."""
+        if self.transform == "log10":
+            return math.log10(value)
+        return value
+
+    def to_physical(self, value: float) -> float:
+        """Sampling/optimization space -> physical value."""
+        if self.transform == "log10":
+            return 10.0**value
+        return value
+
+    @property
+    def sample_bounds(self) -> tuple[float, float]:
+        """The bounds expressed in sampling space."""
+        lo, hi = self.bounds
+        return (self.to_sample_space(lo), self.to_sample_space(hi))
+
+
+def _set_by_path(obj, segments: list[str], value: float):
+    """Return a copy of a pyswap object with the addressed location replaced.
+
+    Works recursively: each level is rebuilt through the object's `update`
+    method so that pydantic validation stays in force.
+    """
+    head, rest = segments[0], segments[1:]
+
+    cell = _CELL_RE.match(head)
+    if cell:
+        if rest:
+            msg = f"table cell must be the final path segment, got {segments}"
+            raise ValueError(msg)
+        column, row = cell.group(1), int(cell.group(2))
+        if not isinstance(obj, pd.DataFrame):
[…]
