### 2026-08-29 eda8238
fix: tolerate models pickled before the SSDI fields existed
Instances restored from pickle archives (e.g. the HDF5 database) lack
the new SSDI attributes and private state; guard attribute access with
getattr and hold the SSDI file-content field names in a module-level
constant instead of a private attribute.
