+++ pyswap/model/model.py
@@ -92,7 +92,7 @@ class ModelBuilder:
             inputs["met"] = self.model.meteorology.met
         if self.model.fixedirrigation.swirgfil == 1:
             inputs["irg"] = self.model.fixedirrigation.irg
-        if self.model.fixedirrigation.swssdi == 1:
+        if getattr(self.model.fixedirrigation, "swssdi", None) == 1:
             inputs["ssdi"] = self.model.fixedirrigation.ssdi
         if self.model.bottomboundary.swbbcfile == 1:
             inputs["bbc"] = self.model.bottomboundary.bbc
@@ -126,7 +126,7 @@ class ModelBuilder:
             self.model.meteorology.write_met(self.tempdir)
         if self.model.fixedirrigation.swirgfil == 1:
             self.model.fixedirrigation.write_irg(self.tempdir)
-        if self.model.fixedirrigation.swssdi == 1:
+        if getattr(self.model.fixedirrigation, "swssdi", None) == 1:
             self.model.fixedirrigation.write_ssdi(self.tempdir)
         if self.model.bottomboundary.swbbcfile == 1:
             self.model.bottomboundary.write_bbc(self.tempdir)
