### 2026-08-29 859fdaa
style: explicit strict in zip
