### 2026-08-29 6f82a03
feat: add one-call sensitivity analysis on SALib
run_morris and run_sobol sample in the parameters' transform space,
evaluate the batch through the parameter core, and return tidy
(output, parameter) DataFrames; multi-output responses are supported
and failed runs raise instead of corrupting the matrices. SALib ships
as the optional 'sa' extra with a lazy import and a clear install hint.
No plotting - that belongs to the analysis notebooks.
