### 2026-08-31 zawadzkim/pySWAP#124 @martinvonk
JOSS Review Comments
Hi @zawadzkim and @markvdbrink,

Thank you for the submission to JOSS (https://github.com/openjournals/joss-reviews/issues/11176) and the great Python package. I think there is really a demand for this package and it is a valuable contribution. I have some comments on the package and the manuscript. The manuscript is overall very solid and I don’t have too many comments on that. The main thing that is missing in my opinion is a good example of a (non standard) use case of the API. I think that can easily be done, while saving some space by getting rid of some redundant text in the manuscript itself. On the package itself I also have some questions, mainly related to some design choices that I find a bit awkward / unidiomatic. I’ve outlined them below.

### Package comments
- Some of the code and design choices feel a bit weird. This might not be my full jurisdiction as reviewer, but quite a lot of the code and design choices. The code claims to be based on FloPy, as a FloPy contributor I have some thoughts about this. FloPy does four things: load models, build models (from different inputs), run models, visualize models. I’m seeing loading and running, but not much of the building and visualization. While I think that would be the interesting and powerful part of coupling SWAP to Python.
- The `py.typed` file is mislocated, and should be in the package source root.
- Tutorial notebooks are on a separate repo, why? Now if you fork / clone the repo you can't run the examples. Also running the examples gives a `FileNotFoundError: [Errno 2] No such file or directory: 'pySWAP/pyswap/libs/WOFOST_crop_parameters/potato.yaml'`
- pySWAP integrates with external databases, such as KNMI for Dutch weather data. Even though this is nice, I think you could also make this optional or remove it altogether. That way you focus on your core requirement that is running SWAP. You can then show in an example notebook how to load data from the KNMI to as input for a SWAP model.
- Maybe it is smarter to split WOFOST and SWAP? I believe you should be able to run SWAP without WOFOST? You could make two Python packages out of it. Same as the KNMI. Make these sort of dependencies external databases that you can couple to SWAP but don’t have to per se if you don’t have to. Even though for WOFOST I can imagine that it’s a stronger coupling.
- Seaborn is a dependency but only used for creating the axes for plotting. I’d suggest to just use Matplotlib. If you really want to use seaborn, then use it to the full extent to make fancy figures. But my thought is that matplotlib would more than suffice here.
- Chardet is a dependency but why? I think you should be able to do without it if youre reading simple ASCII files.
- Please support pandas > 3.0. That way you are future proof. I think this should not be too hard to do.
- pyproj and shapely are dependencies but they are only used to save the model location, which is in itself optional. Consider dropping this and let users figure this out themselves as it is only model metadata.

### Manuscript comments

Using: https://github.com/openjournals/joss-reviews/issues/11176#issuecomment-5358722626

- Line 11: Worth mentioning SWAP's educational GUI as context for why a programmatic Python alternative is beneficial.
- Line 23 separates the European Union and the Netherlands. But the Netherlands is still part of the European Union last time i checked 🙂
- Line 30: What do you mean here? MODFLOW also uses ASCII files for the input and output as do many other hydrological models. There are generally no issues with that.
- Line 35: I think pHYDRUS also deserves a mention. That is a Python package that runs HYDRUS that also simulates 1D unsaturated flow (https://meetingorganizer.copernicus.org/EGU2020/EGU2020-15377.html). Also FloPy supports MODFLOW-USG Tranpsort (https://www.gsienv.com/software/modflow-usg/modflow-usg/) which can simulate 3D variably saturated flow.
- Line 43: “flopy” -> “FloPy”
- Line 43: please check first names in the references. Also I think it is fine to just cite Bakker et al. 2016 and Hughes et al. 2024
- Line 45: I’d line break between: “adoption.” and “Beyond”
- Line 57: but not currently? When will you do this? also as i mentioned, I have no problem with only supporting asci at the moment
- Line 57: there are references for NumPy, Pandas and maybe also pydantic, please use those.
- Line 67: I think it is a bit awkward to run a CLI before running e.g. a normal script / notebook. Can’t you just make some functions that take care of installation the first time you import pyswap?
- Line 70: I have some bad experience with Binder personally. You could consider swapping it out for a GitHub codespace and do everything with GitHub. If not, which is fine obviously, please put an url here to the Binder environment.
- Line 71: On the code example: ps.components.meteorology` feels a bit overly nested and thus overly redundant for importing a Meteorology class. Why not make Meteorology available at the namespace level?
- Line 71: on the code example: it is nice that you can import test cases. But 99.99% of users won’t import a testcase (csv) file. But they want to use their own data. Is a pandas.DataFrame supported with your own data? To Meteorology a metfile keyword argument has to be provided but is a pandas.DataFrame not much more universal here. Instead of weird / native SWAP files / classes?
- Line 71: The code example is in my opinion a bit marginal. I’d like at least an example that solves an actual problem where you show all four steps: preprocessing the data, creating the input for swap, running the swap model, postprocessing the output. E.g.: meteorological time series with precipitation and evaporation, a soil with three different layers (sand, clay, sand), constant head boundary at the bottom, solving the model and showing the results (for instance the wetting front at different time steps). You can make a nice notebook out of this I think that can serve as an example for your documentation as well. You don't have to show all the code in the notebook, but maybe just some of the main steps and results.
- Line 77: Does this HDF5 backend work with SWAP? Or is this only something that works with PySWAP? If the latter is the case? Why would you do this? You already have the ASCII input/output for the SWAP model so you can save the model that way. If you think that’s too large, just zip it. If you want to save the PySWAP model, you can just pickle it which is native python for which you don’t have to use HDF5
- Line 84: Testing: although testing is important I don’t think this is interesting to mention in the manuscript. Or maybe only very briefly alongside the CI/CD.
- Line 88: CI/CD: also important but not very interesting
- Line 94: Are all these bullets listed newly started European projects?
- Line 97: WaterScape is the current project? So mention that first?
- Line 100: “where pyswap will be used to simulate vadose zone dynamics”. I think you mean SWAP here. As PySWAP does not simulate vadose zone dynamics
- Line 102: Maybe condense this paragraph a little bit. Part of it is already mentioned earlier in the statement of need. Also I’m convinced that PySWAP will be useful. You don’t have to mention github stars and number of releases for that.
