### 2026-08-31 #11176 @martinvonk
Hi @rwegener2, @jlarsen-usgs and @zawadzkim ,

I've made some comments on the manuscript and package in: https://github.com/zawadzkim/pySWAP/issues/124.

Currently, I am unable to run the example notebooks, which prevents me from fully testing the package end-to-end. I would appreciate it if the authors could review the comments in the issue, particularly regarding the tutorials and API structure so I can re-test the package once resolved.

Note on availability: I will be on holiday in a few weeks and returning in mid-October. I also have another JOSS review (https://github.com/openjournals/joss-reviews/issues/11173) queued upon my return so my response may be delayed so apologies in advance for any delay.
